+++ src/StateSelector/StateSelector.tsx
@@ -4,17 +4,30 @@
  * SPDX-License-Identifier: LicenseRef-Nordic-4-Clause
  */
 
-import React, { FC, useState } from 'react';
+import React, { FC, useEffect, useRef, useState } from 'react';
 import { Button, ButtonGroup } from 'react-bootstrap';
-import { arrayOf, bool, func, number, string } from 'prop-types';
 
 import './state-selector.scss';
 
-interface Props {
+const useSynchronisationIfChangedFromOutside: <T>(
+    v: T,
+    set: (value: T) => void
+) => void = (externalValue, setInternalValue) => {
+    const previousExternalValue = useRef(externalValue);
+    useEffect(() => {
+        if (previousExternalValue.current !== externalValue) {
+            setInternalValue(externalValue);
+            previousExternalValue.current = externalValue;
+        }
+    });
+};
+
+export interface Props {
     items: string[];
     defaultIndex: number;
     onSelect: (index: number) => void;
     disabled?: boolean;
+    selectedItem?: string;
 }
 
 const StateSelector: FC<Props> = ({
@@ -22,8 +35,11 @@ const StateSelector: FC<Props> = ({
     defaultIndex = 0,
     onSelect,
     disabled = false,
+    selectedItem = items[defaultIndex],
 }) => {
-    const [selected, setSelected] = useState(items[defaultIndex] ?? items[0]);
+    const [selected, setSelected] = useState(selectedItem);
+
+    useSynchronisationIfChangedFromOutside<string>(selectedItem, setSelected);
 
     const selectionButton = (item: string, index: number) => (
         <Button
@@ -45,13 +61,6 @@ const StateSelector: FC<Props> = ({
             {items.map((item, index) => selectionButton(item, index))}
         </ButtonGroup>
     );
-};
-
-StateSelector.propTypes = {
-    items: arrayOf(string.isRequired).isRequired,
-    defaultIndex: number.isRequired,
-    onSelect: func.isRequired,
-    disabled: bool,
 };
 
 export default StateSelector;